move compilerhookplugin class
